Fixed some compilation problems
This changes the API to use double instead of float variables (except for sample buffers). Internally we already use double for floating-point settings so it makes sense to expose this to the world too.

This will cause warnings with applications that call the "get" functions and store the result in a float (not a double) but those should be harmless and can be easily solved by the maintainers of those applications.

The functions that now take "double" instead of "float" arguments, should not cause any warnings or other trouble.