Added a Visual Studio project for the front end
A small number of changes needed to be made to compile it successfully without warnings.

It depends on libsndfile being installed in Program Files (x86). Use the Win32 installer that you can download from http://www.mega-nerd.com/libsndfile/ before attempting to compile this.

The libsndfile.dll file will be automatically copied to the output directory as part of the build.